Migrar de Vercel KV a Firebase
- Reemplazar utils/siteInfo.ts con utils/firebase.ts
- Actualizar _app.tsx, update-site-info.ts y StoreConfiguration.tsx para usar Firebase
- Implementar funciones getSiteInformation, updateSiteInformation y uploadImage en Firebase
- Actualizar useSiteInfo hook para trabajar con Firebase
