+++ package.json
@@ -18,6 +18,7 @@
     "@vercel/kv": "^2.0.0",
     "axios": "^1.4.0",
     "browser-image-compression": "^2.0.2",
+    "firebase": "^10.13.2",
     "framer-motion": "^10.13.1",
     "googleapis": "^122.0.0",
     "mercadopago": "^1.5.17",
